+++ packages/vant-icons/CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Changelog
 
+## 1.8.0
+
+- add qq icon
+- add weibo icon
+- add link-o icon
+- add miniprogram-o icon
+- add wechat-moments icon
+
 ## 1.7.3
 
 - complete https protocol